update: API trips types
