+++ src/api/types/ticket.ts
@@ -1,3 +1,8 @@
+import { ITicket, IPassengerSeat } from '../../types/ticket';
+
+// TODO: Remove this export after migrating rest of the code
+export type { ITicket, IPassengerSeat };
+
 export interface ITicketRequest {
     pnrNumber: string;
 }
@@ -10,20 +15,3 @@ export interface IBookingRequest {
 }
 
 export interface IBookingResponse extends ITicket {}
-
-export interface ITicket {
-    pnrNumber: string;
-    tripId: string;
-    seats: IPassengerSeat[];
-}
-
-export interface IPassengerSeat {
-    seatNumber: number;
-    passenger: IPassenger;
-}
-
-export interface IPassenger {
-    fullName: string;
-    age: number;
-    gender: string;
-}
