+++ src/api/types/trip.ts
@@ -1,3 +1,35 @@
+import { IBusType, ISeat, ISeatType, ITrip } from '../../types';
+import { ILocation } from './location';
+
+// TODO: Remove this export after migrating rest of the code
+export type { IBusType, ISeat, ISeatType, ITrip };
+
+interface ITripExternal {
+    id: number;
+    createdAt: string;
+    updatedAt: string;
+    originId: number;
+    destinationId: number;
+    tripDate: string;
+    departure: string;
+    arrival: string;
+    durationInHours: string;
+    busId: string;
+    busType: string;
+    seatType: string;
+    totalSeats: number;
+    farePerSeat: string;
+    publish: boolean;
+    creatorId: string;
+    origin: ILocation;
+    destination: ILocation;
+    creator: unknown;
+}
+
+interface ITripDetailExternal extends ITripExternal {
+    seats: ISeat[];
+}
+
 export interface ITripsQueryRequest {
     originId: string;
     destinationId: string;
@@ -11,6 +43,16 @@ export interface ITripsQueryRequest {
     passengerCount?: number;
 }
 
+export interface ITripsQueryResponse {
+    trips: ITripExternal[];
+}
+
+export interface ITripDetailRequest {
+    id: string;
+}
+
+export interface ITripDetailResponse extends ITripDetailExternal {}
+
 export enum ITripsSortKey {
     DEPARTURE_TIMESTAMP = 'departureTimestamp',
     ARRIVAL_TIMESTAMP = 'arrivalTimestamp',
@@ -20,49 +62,4 @@ export enum ITripsSortKey {
 export enum ISortOrder {
     ASC = 'ASC',
     DESC = 'DESC',
-}
-
-export interface ITripsQueryResponse {
-    trips: ITrip[];
-}
-
-export interface ITripDetailRequest {
-    id: string;
-}
-
-export interface ITripDetailResponse extends ITrip {
-    seats: ISeat[];
-}
-
-export interface ITrip {
-    id: string;
-    origin: string;
-    destination: string;
-    departureTimestamp: string;
-    arrivalTimestamp: string;
-    seatType: ISeatType;
-    busType: IBusType;
-    farePerSeat: number;
-    availableSeats: number;
-    totalSeats: number;
-}
-
-export interface ISeat {
-    seatNumber: number;
-    status: ISeatStatus;
-}
-
-export enum ISeatType {
-    SLEEPER = 'SLEEPER',
-    SEATER = 'SEATER',
-}
-
-export enum IBusType {
-    AC = 'AC',
-    NON_AC = 'NON_AC',
-}
-
-export enum ISeatStatus {
-    BOOKED = 'BOOKED',
-    AVAILABLE = 'AVAILABLE',
 }
